Task1_IMDB_WIKI_Age_Detection/requirements.txt
numpy
pandas
matplotlib
scikit-learn
tensorflow
keras
opencv-python
tqdm
jupyter
Pillow